+++ app/services/ai_credential_validation.rb
@@ -9,7 +9,6 @@ class AiCredentialValidation
     run.succeed! do |credential|
       credential.update!(active: true, available_models: models, models_refreshed_at: Time.current,
                          last_validated_at: Time.current, last_error: nil)
-      credential.active_operation_run(:models_refresh)&.supersede!
     end
   end